advance login process #2

+++ src/main/kotlin/noh/jinil/boot/domain/response/ResponseData.kt
@@ -2,11 +2,11 @@ package noh.jinil.boot.domain.response
 
 class ResponseData<T> {
     var message: String? = null
-    var code: Long? = null
+    var code: Int? = null
     var data: T? = null
 
     companion object {
-        private fun <T> create(dataT: T?, resCode: Long?, resMessage: String?): ResponseData<T> {
+        private fun <T> create(dataT: T?, resCode: Int?, resMessage: String?): ResponseData<T> {
             return ResponseData<T>().apply {
                 code = resCode
                 message = resMessage
@@ -18,11 +18,11 @@ class ResponseData<T> {
             return create(dataT, ResponseCode.OK, "")
         }
 
-        fun <T> createFailed(resCode: Long?, resMessage: String?): ResponseData<T> {
+        fun <T> createFailed(resCode: Int?, resMessage: String?): ResponseData<T> {
             return create(null, resCode, resMessage)
         }
 
-        fun <T> createExtra(resCode: Long?, resMessage: String?): ResponseData<T> {
+        fun <T> createExtra(resCode: Int?, resMessage: String?): ResponseData<T> {
             return create(null, resCode, resMessage)
         }
     }
